medical-claims-anomaly-detection/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
joblib>=1.3
python-dateutil>=2.8
matplotlib>=3.7
